+++ Picadito-de-Quidditch/src/main/Escoba.java
@@ -1,5 +1,5 @@
 package main;
 
 public interface Escoba {
-    Integer velocidadDeEscoba();
+    Double velocidadDeEscoba();
 }
