+++ Picadito-de-Quidditch/src/main/Golpeador.java
@@ -4,13 +4,13 @@ public class Golpeador extends Jugador {
     private Integer fuerza;
     private Integer punteria;
 
-    Golpeador(Integer fuerza, Integer punteria, Integer peso, Escoba escoba){
+    Golpeador(Integer fuerza, Integer punteria, Double peso, Escoba escoba){
         super(peso, escoba);
         this.fuerza=fuerza;
         this.punteria=punteria;
     }
 
-    public Integer habilidad(){
+    public Double habilidad(){
         return super.habilidad()+this.punteria+this.fuerza;
     }
 }
